Support nested fields
Includes support for better labeling of oneOf/anyOf.

The general approach is copied over from cloudflare's recently
updated in-house theme.

This involved piping some fields through to figure out the
depth and location of the nested things.  The change for this
to definition.js has no immediate application- in the internal
cloudflare theme it's used for analytics.  But it seemed to fit
logically to make those extra props available there so I went
ahead and made that consistent.